Se añade cliente feign a movies

+++ movies/src/main/java/edu/co/poli/movies/service/MovieServiceImpl.java
@@ -1,17 +1,25 @@
 package edu.co.poli.movies.service;
 
+import edu.co.poli.movies.clientFeign.BookingsClient;
+import edu.co.poli.movies.clientFeign.ShowtimesClient;
+import edu.co.poli.movies.model.Booking;
+import edu.co.poli.movies.model.Showtime;
 import edu.co.poli.movies.persistence.entity.Movie;
 import edu.co.poli.movies.persistence.repository.MovieRepository;
 import lombok.RequiredArgsConstructor;
+import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
 public class MovieServiceImpl implements MovieService{
 
     private final MovieRepository movieRepository;
+    private final BookingsClient bookingsClient;
+    private final ShowtimesClient showtimesClient;
 
 
     @Override
@@ -21,9 +29,18 @@ public class MovieServiceImpl implements MovieService{
     }
 
     @Override
-    public void delete(Movie movie) {
-        movieRepository.delete(movie);
-
+    public String delete(Movie movie) {
+        List<Booking> booking= bookingsClient.findAll();
+        boolean existBooking = booking.stream()
+                .anyMatch(bookingItem -> bookingItem.getMovies().contains(movie.getId()));
+        List<Showtime> showtime = showtimesClient.findAll();
+        boolean existShowtime = showtime.stream()
+                .anyMatch(showtimeItem -> showtimeItem.getMovies().contains(movie.getId()));
+        if(!existBooking && !existShowtime){
+            movieRepository.delete(movie);
+            return "Eliminado correctamente";
+        }
+        return "Existe una programacion o reserva con esta pelicula";
     }
 
     @Override
